prayer_code_app/requirements.txt
flask==2.0.1
numpy==1.21.2
Pillow==8.3.2
scipy==1.7.1
gunicorn==20.1.0